refactor(gui): the shell's chrome moves to its own file (#678)
`SettingsView` came out of the fix batch at 349 lines against a
hard ceiling of 350 — legal, and one comment away from not
being. The chrome — header, banners, content, and the save
surface in whichever form the band gives it — is the part that
reads whole on its own, and the routing plus the state stay
behind.

Three guards followed their subject rather than being widened:
the surfacing needles for the mounts that moved, and the page
token's render site.
